Move toward long long, as that is consistently 8 bytes.

+++ src/ebprot.h
@@ -370,7 +370,7 @@ int getche(void) ; //@
 int getch(void) ; //@
 #endif // #ifndef _INC_CONIO
 char getLetter(const char *s) ;
-char *getFileName(int msg, const char *defname, bool isnew, bool ws, size_t filesize);
+char *getFileName(int msg, const char *defname, bool isnew, bool ws, unsigned long long int filesize);
 int shellProtectLength(const char *s);
 void shellProtect(char *t, const char *s);
 char *dirSuffixWindow(int n, const Window *w);
